🩹 기준문서 삭제 case 에러 핸들링 추가

+++ app/services/standard/vector_delete.py
@@ -9,14 +9,16 @@ from app.common.exception.error_code import ErrorCode
 
 def delete_by_standard_id(standard_id: int):
   filter_condition = models.Filter(
-      must=[
-        models.FieldCondition(
-            key="standard_id",
-            match=models.MatchValue(value=standard_id)
-        )
-      ]
+      must=[models.FieldCondition(key="standard_id",
+                                  match=models.MatchValue(value=standard_id))]
   )
 
+  existing_data = qdrant_db_client.scroll(collection_name="standard",
+                                          scroll_filter=filter_condition,
+                                          limit=1)
+
+  if len(existing_data[0]) == 0:
+    raise StandardException(ErrorCode.NOT_FOUND)
   try:
     qdrant_db_client.delete(
         collection_name="standard",
